Add Handlebars validation helpers and integrate into Dashlet components
