+++ apps/app/src/features/dashboard/dashlets/text_card/dashlet.settings.tsx
@@ -3,64 +3,18 @@
 import { useState, useRef } from "react";
 import { Button, Label, TextInput, Textarea, ToggleSwitch } from "flowbite-react";
 import { createPortal } from "react-dom";
-import Handlebars from "handlebars";
 import { twMerge } from "tailwind-merge";
-import type { DashletSettingsProps,DataProviderEntry } from "../types";
+import type { DashletSettingsProps, DataProviderEntry } from "../types";
 import type { DashletConfig, TextAlign } from "./dashlet";
 import { tr } from "@/features/i18n/tr.service";
 import AbsoluteModal from "@/features/common/components/absolute-modal/absolute-modal";
-import { SettingsSelectField } from "../common";
+import { SettingsSelectField, getHandlebarsStatus, getFlowbiteColor } from "../common";
 
 // ============================================================================
 // Types
 // ============================================================================
 
 type SettingsTab = "visualization" | "data";
-
-// ============================================================================
-// Handlebars validation helpers
-// ============================================================================
-
-type HandlebarsStatus = "valid" | "invalid" | "none";
-
-function findHandlebarsExpressions(text: string): string[] {
-  const results: string[] = [];
-  let start = text.indexOf("{{");
-  while (start !== -1) {
-    const end = text.indexOf("}}", start + 2);
-    if (end === -1) break;
-    results.push(text.substring(start, end + 2));
-    start = text.indexOf("{{", end + 2);
-  }
-  return results;
-}
-
-function getHandlebarsStatus(text: string): HandlebarsStatus {
-  const matches = findHandlebarsExpressions(text);
-  if (matches.length === 0) return "none";
-  for (const match of matches) {
-    const inner = match.slice(2, -2).trim();
-    if (
-      !inner ||
-      inner.endsWith(".") ||
-      inner.startsWith(".") ||
-      /\.\./.test(inner) ||
-      !/^[\w\s.#/^>@!-]+$/.test(inner)
-    ) return "invalid";
-  }
-  try {
-    Handlebars.compile(text);
-    return "valid";
-  } catch {
-    return "invalid";
-  }
-}
-
-function getFlowbiteColor(s: HandlebarsStatus): "gray" | "success" | "failure" {
-  if (s === "valid") return "success";
-  if (s === "invalid") return "failure";
-  return "gray";
-}
 
 // ============================================================================
 // Settings Component
